added functionality for string.gsub in invitation whisper
I wanted to have an invitation message that advocates for the use of inverse phrases in the /layer channel something akin to this:

Inviting you to layer 1. If you are UNHAPPY with this layer, decline my invite and type in /layer channel like "layer not 1"

strong.format doesn't natively allow you to use multiple %s values in the string so I created a helper function to also allow the use of string.gsub to look for the characters: {layer}.

This allows you to use the layer number multiple times in a whisper.